update deleteProduct in controller

+++ backend/controllers/product.controller.js
@@ -51,6 +51,10 @@ export const updateProduct = async(req, res) => {
 export const deleteProduct = async(req, res) => {
     try {
        const {id} = req.params;
+
+       if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(400).json({success: false, message: "Invalid produc id"})
+       };
        
        await Product.findOneAndDelete(id);
        res.status(200).json({success: true, message: 'Deletes product successfully'})
